+++ lib/core/constants/bools.dart
@@ -1,0 +1,6 @@
+class Booleans{
+  ///app booleans
+  static const inDebugMode = true;
+  static const superPrintMode = true;
+  static const inBeta = false;
+}