+++ lib/core/constants/images.dart
@@ -1,0 +1,3 @@
+class AppImages{
+  static const logoGIF = 'assets/niveus.gif';
+}